+++ staircase/docstrings/stairs_module.py
@@ -113,9 +113,7 @@ hist_from_ecdf_example = """
         >>> for ax, title, stair_instance in zip(axes, ("s1", "s1 ecdf"), (s1, s1_ecdf_stairs)):
         ...     stair_instance.plot(ax)
         ...     ax.set_title(title)
-
-    .. plot::
-        :context: close-figs
+        >>> plt.show() 
         
         >>> sc.hist_from_ecdf(s1_ecdf_stairs)
         [-1, 0)    0.25
@@ -127,4 +125,58 @@ hist_from_ecdf_example = """
         [-1, 1)    0.5
         [1, 3)     0.5
         dtype: float64
+"""
+
+cov_example = """
+    Examples
+    --------
+        
+    .. plot::
+        :context: close-figs
+        
+        >>> import staircase as sc
+        >>> pd.Series([s1, s2, s1+s2])
+        0    <staircase.Stairs, id=2452772382088, dates=False>
+        1    <staircase.Stairs, id=2452772381320, dates=False>
+        2    <staircase.Stairs, id=2452772893512, dates=False>
+        dtype: object
+        
+        >>> sc.cov(pd.Series([s1, s2, s1+s2], index=['s1', 's2', 's1+s2']))
+                      s1         s2      s1+s2
+        s1      0.687500   0.140496   0.652893
+        s2      0.140496   0.471074   0.611570
+        s1+ s2  0.652893   0.611570   1.264463
+        
+        >>> sc.cov([s1, s2, s1+s2])
+                  0          1          2 
+        0  0.687500   0.140496   0.652893
+        1  0.140496   0.471074   0.611570
+        2  0.652893   0.611570   1.264463        
+"""
+
+corr_example = """
+    Examples
+    --------
+        
+    .. plot::
+        :context: close-figs
+        
+        >>> import staircase as sc
+        >>> pd.Series([s1, s2, s1+s2])
+        0    <staircase.Stairs, id=2452772382088, dates=False>
+        1    <staircase.Stairs, id=2452772381320, dates=False>
+        2    <staircase.Stairs, id=2452772893512, dates=False>
+        dtype: object
+        
+        >>> sc.corr(pd.Series([s1, s2, s1+s2], index=['s1', 's2', 's1+s2']))
+                      s1         s2      s1+s2
+        s1      1.000000   0.246878   0.700249
+        s2      0.246878   1.000000   0.792407
+        s1+ s2  0.700249   0.792407   1.000000
+        
+        >>> sc.corr([s1, s2, s1+s2])
+                  0          1          2 
+        0  1.000000   0.246878   0.700249
+        1  0.246878   1.000000   0.792407
+        2  0.700249   0.792407   1.000000       
 """